ticket-env/support_ticket_env/server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.115.0
uvicorn>=0.24.0
sqlalchemy
aiosqlite



